Added gradient field option for showing gravity
There are now two views to show the gravity field in the simulator. 1) As
a vector field with lines. 2) As a gradient field (new option).

The brightness of a point on the gradient field depends on the net
gravitational force. So the point between objects where the net force is 0
should be dark.