Added asteroid gravity shape type
Right-click to add an asteroid.

Makes a red asteroid which is a polygon with equidistant angles between
the vertices (relative to the center). The distance of the vertices from
the center is also somewhat randomized so they don't all come out as
regular polygons.

The next step is to add collisions for the asteroids, hopefully.