Added collision momentum and particles
When an asteroid breaks apart the pieces will now get a portion of the
momentum of the other object the asteroid collided with.

When a point collides with an asteroid CollisionParticles are now created.
They fade from white to yellow to red to black before being removed. They
still need work to fade at different speeds and get created with different
colors.

Collisions still need work to prevent checking a point that has already
collided with something and is scheduled to be deleted.